Finally got the teapot to display
It wasn't lighting, it turns out it was camera position. After writing a
little Task, to randomly change the camera position, I found a location
that worked. Then I realised you could just have the camera lookAt() the
teapot.

However, this is very much NOT the loading screen, so I'm going to move
it next commit.