Started coding main-view.py
The main view is the view with the spaceships. More verbosely, it deals
with the primary "in space" view, where you swivel the camera around
the fixed point, and the one in which you enter to actually see what is
going on.

It'll also probably be the most graphically intensive part of the game
considering it displays all the explosions and ships, and jazz.