Added FocusManager, a utility class for focusing
"Focusing" or having focus on multiple items means that the camera
pivots and swivels around a central point, and that point is the average
of all objects which are focused. The FocusManager is a set which you
add items to be focused to, and will tell the camera where to look.