Changed LoadingScreen -> TeapotScreen, and cleaned
Renamed the class, and also modified it so all the initalisation logic
only ran if it was called as a script. It still can be called in one go
with .setup(), but if we want to use TeapotScreen as the premade objects
to debug, then we need to make it less enthusiastic.

As in, not changing the background, setting up the spooky camera
movement, and so on, unless we ask for it.